Start down the road of drastic simplification
I like the philosophical approach to the original NANO theme and the
basic idea of the color usage. However, the actual implementation
seems a bit overly clever for my needs.

I'm working towards follow a pattern more like what I've found in the
'Almost Mono' themes
https://github.com/cryon/almost-mono-themes/. Nothing too fancy but
supports multiple versions of the themes in a straightforward manner.